+++ R/class-workbook.R
@@ -3780,6 +3780,7 @@ wbWorkbook <- R6::R6Class(
       if (length(cols) == 0L) {
         return(self)
       }
+      assert_class(hidden, "logical", arg_nm = "hidden")
 
       cols <- col2int(cols)
 
@@ -3791,8 +3792,25 @@ wbWorkbook <- R6::R6Class(
         stop("hidden argument is longer than cols.")
       }
 
+      compatible_length <- length(cols) %% length(widths) == 0
+
+      if (!compatible_length) {
+        # needed because rep(c(1, 2 ), length.out = 3) is successful,
+        # but not clear if this is what the user wanted.
+        stop("`cols` and `widths` don't have compatible lengths.\n",
+             "`cols` has length ", length(cols), " while ",
+             "`widths` has length ", length(widths), ".")
+      }
+
       if (length(widths) < length(cols)) {
         widths <- rep(widths, length.out = length(cols))
+      }
+      compatible_length <- length(cols) %% length(hidden) == 0
+
+      if (!compatible_length) {
+        stop("`cols` and `hidden` must have compatible lengths.\n",
+             "`cols` has length ", length(cols), " while ",
+             "`hidden` has length ", length(hidden), ".")
       }
 
       if (length(hidden) < length(cols)) {