+++ NEWS.md
@@ -17,12 +17,12 @@ wb_dims(x = mtcars, cols = "non-existent-col")
 # Now errors
 ```
 
+* `wb_set_col_widths()` is more strict about its arguments. If you provide `cols`, `widths`, or `hidden` don't have appropriate length, it will throw an error.
+
 
 ***************************************************************************
 
 # openxlsx2 1.5
-
-(This was updated post release.)
 
 ## New features